Added fetchAuthCode() in new file
